feat(map): use new memory layout
 * instead of one block of memory which makes it necessary to reallocate
   memory (and copy data) if more space is needed the BucketArray
   contains multiple bucketBuffers which can contain a fix size of
   buckets if more space is needed a new bucketBuffer is allocate.
   No need of reallocting and copy data.
 * this increased the performance about 30 %
 * renamed BucketArray to BucketBufferArray
 * add some utilities like toString
 * use lomg as hashcode for less colisions
 * create a bunch of tests also for the IO for the BucketBufferArray

 related to camunda-zeebe/zeebe#358

+++ zb-map/src/main/java/io/zeebe/map/KeyHandler.java
@@ -19,7 +19,7 @@ public interface KeyHandler
 {
     void setKeyLength(int keyLength);
 
-    int keyHashCode();
+    long keyHashCode();
 
     boolean keyEquals(long keyAddr);
