+++ zb-map/src/main/java/io/zeebe/map/types/LongKeyHandler.java
@@ -34,7 +34,7 @@ public class LongKeyHandler implements KeyHandler
     }
 
     @Override
-    public int keyHashCode()
+    public long keyHashCode()
     {
         return Long.hashCode(theKey);
     }
